Вынесла state из SignIn в хок with-auth-information
